+++ python-lib/dku_model_fairness_report/__init__.py
@@ -1,3 +1,1 @@
-# coding: utf-8
-from dku_model_fairness_report.fairness_metric_report import ModelFairnessMetric
-from dku_model_fairness_report.constants import DkuFairnessConstants
+from dku_model_fairness_report.fairness_metric_report import ModelFairnessMetricReport, ModelFairnessMetric
