+++ python-lib/dku_webapp/constants.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+class DkuWebappConstants(object):
+
+    NAME = 'name'
+    SIZE = 'size'
+    PREDICTION = 'prediction'
