+++ resource/templates/modal.html
@@ -6,16 +6,7 @@
     <div class="dku-modal-body" ng-show="modal.msg" ng-class="{'dku-modal-body-error': modal.type === 'error'}">
         {{modal.msg}}
     </div>
-    <form name="promptForm" class="dku-modal-body" ng-show="modal.type === 'prompt'" ng-submit="promptForm.$valid && modal.confirmAction(); removeModal()">
-        <div>
-            {{modal.inputLabel}}
-        </div>
-        <input focus-here ng-model="modal.promptResult" required>
-    </form>
-    <div class="dku-modal-footer">
-        <div class="dku-button-modal-div">
-            <button type="button" class="button dku-button-modal" ng-if="modal.type === 'prompt' || modal.type === 'confirm'" ng-click="removeModal()">Cancel</button>
-            <button type="button" class="button dku-button-modal" ng-click="(modal.confirmAction && modal.confirmAction()); removeModal();" ng-disabled="modal.type == 'prompt' && promptForm.$invalid">Ok</button>
-        </div>
+    <div>
+            <button type="button" class="button dku-button-modal dku-button-modal-div" ng-click="(modal.confirmAction && modal.confirmAction()); removeModal();" ng-disabled="modal.type == 'prompt' && promptForm.$invalid">Ok</button>
     </div>
 </div>
