+++ webapps/model-fairness-view/body.html
@@ -30,7 +30,7 @@ with reference group
     </select>
     <div style="display: inline;">
         <a class="dku-btn dku-btn-primary notrunning-state" id="run-button" type="button" ng-click="runAnalysis()" style="color: #FFFFFF;">COMPUTE</a>
-        <span class="running-state" style="display: none;">Running ...</span>
+        <span class="running-state" style="display: none;">Computing ...</span>
     </div>
 </div>
 </div>
@@ -45,7 +45,7 @@ with reference group
 
 <div class="landing-page center" style="padding-top:16px; padding-bottom:16px ">
 <div class="ultralarge-title-sb">
-    Model Fairness
+    Model Fairness Report
 </div>
 <div class="grand-title-r">
     Select column and advantageous outcome on which you want to compute fairness metrics
