+++ webapps/model-fairness-view/style.css
@@ -320,11 +320,6 @@ table{
     opacity: 0.3;
 }
 
-
-
-.info-box-2 {
-}
-
 .info-box > .info-box-item {
     background: #ffffff;
     box-shadow: 0px 0px 2px 1px rgba(34, 34, 34, 0.15);
@@ -343,7 +338,6 @@ table{
   flex-basis: 100%;
   height: 0;
 }
-
 
 canvas {
   -moz-user-select: none;
@@ -370,8 +364,24 @@ canvas {
 }
 
 .dku-button-modal {
-    font-weight: 500;
+    font-weight: 450;
+    font-family: 'SourceSansPro';
+    text-transform: uppercase;
     font-size: 14px;
+    padding: 3px 8px;
+    line-height: 1.4;
+    background-image: inherit;
+    box-shadow: none;
+    text-shadow: none;
+    box-sizing: border-box;
+    outline: 0;
+    cursor: pointer;
+    color: black;
+    border: 1px solid #cccccc;
+    height: 26px;
+    border-color: transparent;
+    background-color: transparent;
+    margin: 0;
 }
 
 .dku-button-modal:disabled {
@@ -391,7 +401,8 @@ canvas {
 .dku-modal {
     max-height: 360px;
     width: 500px;
-    padding: 20px;
+    padding: 16px;
+    padding-bottom: 12px;
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
     background-color: var(--white);
 }
@@ -406,6 +417,7 @@ canvas {
     white-space: pre-line;
     text-align: justify;
     font-size: 14px;
+
 }
 
 .dku-modal-body-error {
